Restructured the project structure and Wellness App package
