+++ app/src/main/java/com/rajit/learning_compose/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
 import androidx.compose.material3.windowsizeclass.calculateWindowSizeClass
+import com.rajit.learning_compose.wellnessapp.WaterCounterApp
 
 class MainActivity : ComponentActivity() {
